carbon: show retired and redeemed cards for pools
