+++ carbon/components/cards/tokenDetails/TokenDistributionOfMethodologiesCard/index.tsx
@@ -13,11 +13,10 @@ import { queryAggregatedCreditsByPoolAndMethodology } from "lib/charts/queries";
 export default function TokenDistributionOfMethodologiesCard(
   props: CardProps & TokenDetailsProps
 ) {
-  // No methodologies card for retired credits on particular pools
-  /*
+  // No methodologies distribution card for retired credits on particular pools
   if (props.pool != "all" && props.status == "retired") {
     return <></>;
-  }*/
+  }
   const chart = (
     /* @ts-expect-error async Server component */
     <TokenDistributionOfMethodologiesChart {...props} />
