+++ carbon/components/cards/tokenDetails/TokenDistributionOfVintageCard/index.tsx
@@ -13,11 +13,10 @@ import { queryAggregatedCreditsByPoolAndVintage } from "lib/charts/queries";
 export default function TokenDistributionOfVintageCard(
   props: CardProps & TokenDetailsProps
 ) {
-  // No vintage card for retired credits on particular pools
-  /*
+  // No vintage distribution card for retired credits on particular pools
   if (props.pool != "all" && props.status == "retired") {
     return <></>;
-  }*/
+  }
   const chart = (
     /* @ts-expect-error async Server component */
     <TokenDistributionOfVintageChart {...props} />
