+++ carbon/components/cards/tokenDetails/TokenOriginsCard/index.tsx
@@ -7,6 +7,10 @@ import { creditsQueryParamsFromProps } from "lib/charts/aggregators/getAggregate
 import layout from "theme/layout.module.scss";
 
 export default function TokenOriginsCard(props: CardProps & TokenDetailsProps) {
+  // No token origin card for retired credits on particular pools
+  if (props.pool != "all" && props.status == "retired") {
+    return <></>;
+  }
   const params = creditsQueryParamsFromProps(props);
   const chart = (
     /* @ts-expect-error async Server component */
