+++ carbon/components/cards/tokenDetails/helpers.ts
@@ -95,7 +95,6 @@ export function getChartConfiguration(props: TokenDetailsProps) {
 export function getCreditsQueryConfiguration(
   props: TokenDetailsProps
 ): DailyCreditsQueryConfiguration {
-  const freq = props.since == "lifetime" ? "monthly" : "daily";
   const dateField = statusToDateField(props.status);
   const params = creditsQueryParamsFromProps(props);
